+++ web/Controllers/PlantsController.cs
@@ -105,6 +105,7 @@ namespace web.Controllers
         // GET: Plants/Details/5
         public async Task<IActionResult> Details(int? id)
         {
+            var currentUser = await _usermanager.GetUserAsync(User);
             if (id == null)
             {
                 return NotFound();
@@ -113,6 +114,7 @@ namespace web.Controllers
             var plant = await _context.Plants
                 .Include(p => p.Category)
                 .Include(p => p.Location)
+                .Where(p => p.User == currentUser)
                 .FirstOrDefaultAsync(m => m.PlantID == id);
             if (plant == null)
             {
@@ -173,7 +175,7 @@ namespace web.Controllers
             }
 
             var plant = await _context.Plants.FindAsync(id);
-            if (plant == null)
+            if (plant == null || plant.User != currentUser)
             {
                 return NotFound();
             }
@@ -189,7 +191,8 @@ namespace web.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, [Bind("PlantID,Name,Description,Note,image,DaysBetweenWatering,LastWateredDate,CategoryID,LocationID")] Plant plant)
         {
-            if (id != plant.PlantID)
+            var currentUser = await _usermanager.GetUserAsync(User);
+            if (id != plant.PlantID || plant.User != currentUser)
             {
                 return NotFound();
             }
@@ -229,10 +232,12 @@ namespace web.Controllers
                 return NotFound();
             }
 
+            var currentUser = await _usermanager.GetUserAsync(User);
+
             var plant = await _context.Plants
                 .Include(p => p.Category)
                 .Include(p => p.Location)
-                .FirstOrDefaultAsync(m => m.PlantID == id);
+                .FirstOrDefaultAsync(m => m.PlantID == id && m.User == currentUser);
             if (plant == null)
             {
                 return NotFound();
@@ -247,6 +252,11 @@ namespace web.Controllers
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var plant = await _context.Plants.FindAsync(id);
+            var currentUser = await _usermanager.GetUserAsync(User);
+             if (plant.User != currentUser)
+            {
+                return NotFound();
+            }
             _context.Plants.Remove(plant);
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
